fix(v038): harden Tier B threshold validation + correct chair-threshold docs
Deep-review follow-ups on PR 1 (per-disposition `threshold` + `chair`):

- Reject a `threshold` on a non-open-floor disposition. The salience bid only
  runs for open-floor speakers (participant/chair/legacy `always`, all
  normalizing to RespondAlways); a threshold on addressed/observer/
  when_mentioned/never — or on a member that omits `respond` — is a silent
  no-op. New ErrThresholdNotApplicable; loud-fail in Validate (a cross-field
  invariant the JSON schema cannot express, so loader-only).
- Reject a non-finite `.nan` threshold. NaN slips past the `< 0 || > 1` range
  check (every NaN comparison is false) and the schema's numeric bound misses
  it too, so a NaN salience bar would reach the bid where `score >= NaN` is
  unconditionally false. Added an explicit math.IsNaN guard (`±Inf` is already
  caught by the range bound).
- Fix the self-contradicting DefaultChairThreshold rationale: the old comment
  claimed "zero reads as unset → biases to silence", contradicting the
  nil-vs-&0.0 tri-state where &0.0 is an explicit bar every score clears
  (maximally talkative). Rewrote: nil = bias-to-silence; &0.0 would pile onto
  every message; 0.15 is low-but-nonzero so the chair speaks readily while
  respecting a minimal floor.
- Make the persistence/wire gap explicit (docs only; plumbing is PR 2 scope).
  The threshold lives only on the in-memory Config from LoadConfig —
  ReconcileConfig drops everything but ID + policy, and it is absent from the
  ChannelMessageEvent wire the Python gate/bid consume. Corrected the
  overclaiming comments on MemberConfig.Threshold, RespondChair, and the
  Python POLICY_CHAIR alias so PR 2 is not planned on a false premise.

Extract the threshold loader tests into config_threshold_test.go to keep both
files under the 500-line cap.

TDD (red → green): TestLoadConfig_RejectsNaNThreshold,
TestLoadConfig_RejectsThresholdOnNonOpenFloorDisposition (5 sub-cases),
TestLoadConfig_AcceptsThresholdOnLegacyAlways.

go test ./internal/channels/ + go vet + go build ./... green; Python gate +
schema suites green; agents/validate.py config/ green.
