+++ FILEMAP.md
@@ -1,7 +1,7 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-06-07  
-> **1332** tracked files
+> **1333** tracked files
 
 ## Summary
 
@@ -9,7 +9,7 @@
 |---------------------|-------|
 | `docs/` | 325 |
 | `tests/` | 292 |
-| `internal/` | 285 |
+| `internal/` | 286 |
 | `agents/` | 217 |
 | `web/` | 41 |
 | `prompts/` | 30 |
@@ -30,7 +30,7 @@
 |-----------|-------|
 | `.py` | 524 |
 | `.md` | 371 |
-| `.go` | 298 |
+| `.go` | 299 |
 | `.yaml` | 37 |
 | `.js` | 24 |
 | `.rs` | 22 |
@@ -750,6 +750,7 @@
 │   │   ├── channels.go
 │   │   ├── config.go
 │   │   ├── config_test.go
+│   │   ├── config_threshold_test.go
 │   │   ├── epoch_override.go
 │   │   ├── fanout.go
 │   │   ├── fanout_floor_test.go
